+++ linopy/solvers.py
@@ -119,7 +119,7 @@ def run_cbc(
     and constraint dual values. For more information on the solver
     options, run 'cbc' in your shell
     """
-    if io_api is not None and (io_api != "lp"):
+    if io_api is not None and io_api not in ["lp", "mps"]:
         logger.warning(
             f"IO setting '{io_api}' not available for cbc solver. "
             "Falling back to `lp`."
@@ -209,11 +209,14 @@ def run_glpk(
     """
     Solve a linear problem using the glpk solver.
 
-    This function reads the linear problem file and passes it to the glpk
-    solver. If the solution is successful it returns variable solutions and
+    This function reads the linear problem file and passes it to the
+    glpk
+    solver. If the solution is successful it returns variable solutions
+    and
     constraint dual values.
 
-    For more information on the glpk solver options:
+    For more information on the glpk solver options, see
+
     https://kam.mff.cuni.cz/~elias/glpk.pdf
     """
     CONDITION_MAP = {
@@ -221,16 +224,17 @@ def run_glpk(
         "undefined": "infeasible_or_unbounded",
     }
 
-    if io_api is not None and (io_api != "lp"):
+    if io_api is not None and io_api not in ["lp", "mps"]:
         logger.warning(
             f"IO setting '{io_api}' not available for glpk solver. "
             "Falling back to `lp`."
         )
 
     problem_fn = model.to_file(problem_fn)
+    suffix = problem_fn.suffix[1:]
 
     # TODO use --nopresol argument for non-optimal solution output
-    command = f"glpsol --lp {problem_fn} --output {solution_fn} "
+    command = f"glpsol --{suffix} {problem_fn} --output {solution_fn} "
     if log_fn is not None:
         command += f"--log {log_fn} "
     if warmstart_fn:
@@ -339,7 +343,7 @@ def run_highs(
     if warmstart_fn:
         logger.warning("Warmstart not available with HiGHS solver. Ignore argument.")
 
-    if io_api is None or (io_api == "lp"):
+    if io_api is None or io_api in ["lp", "mps"]:
         model.to_file(problem_fn)
         h = highspy.Highs()
         h.readModel(maybe_convert_path(problem_fn))
@@ -347,7 +351,7 @@ def run_highs(
         h = model.to_highspy()
     else:
         raise ValueError(
-            "Keyword argument `io_api` has to be one of `lp`, `direct` or None"
+            "Keyword argument `io_api` has to be one of `lp`, `mps`, `direct` or None"
         )
 
     if log_fn is None:
@@ -371,14 +375,14 @@ def run_highs(
 
         if io_api == "direct":
             sol = pd.Series(solution.col_value, model.matrices.vlabels, dtype=float)
-            dual = pd.Series(solution.row_value, model.matrices.clabels, dtype=float)
+            dual = pd.Series(solution.row_dual, model.matrices.clabels, dtype=float)
         else:
             sol = pd.Series(solution.col_value, h.getLp().col_names_, dtype=float).pipe(
                 set_int_index
             )
-            dual = pd.Series(
-                solution.row_value, h.getLp().row_names_, dtype=float
-            ).pipe(set_int_index)
+            dual = pd.Series(solution.row_dual, h.getLp().row_names_, dtype=float).pipe(
+                set_int_index
+            )
 
         return Solution(sol, dual, objective)
 
@@ -411,7 +415,7 @@ def run_cplex(
     """
     CONDITION_MAP = {"integer optimal solution": "optimal"}
 
-    if io_api is not None and (io_api != "lp"):
+    if io_api is not None and io_api not in ["lp", "mps"]:
         logger.warning(
             f"IO setting '{io_api}' not available for cplex solver. "
             "Falling back to `lp`."
@@ -530,7 +534,7 @@ def run_gurobi(
     warmstart_fn = maybe_convert_path(warmstart_fn)
     basis_fn = maybe_convert_path(basis_fn)
 
-    if io_api is None or (io_api == "lp"):
+    if io_api is None or io_api in ["lp", "mps"]:
         problem_fn = model.to_file(problem_fn)
         problem_fn = maybe_convert_path(problem_fn)
         m = gurobipy.read(problem_fn)
@@ -539,7 +543,7 @@ def run_gurobi(
         m = model.to_gurobipy()
     else:
         raise ValueError(
-            "Keyword argument `io_api` has to be one of `lp`, `direct` or None"
+            "Keyword argument `io_api` has to be one of `lp`, `mps`, `direct` or None"
         )
 
     if solver_options is not None:
@@ -603,7 +607,7 @@ def run_xpress(
     variable solutions and constraint dual values. The xpress module
     must be installed for using this function.
 
-    For more information on solver options:
+    For more information on solver options, see
     https://www.fico.com/fico-xpress-optimization/docs/latest/solver/GUID-ACD7E60C-7852-36B7-A78A-CED0EA291CDD.html
     """
     CONDITION_MAP = {
@@ -616,7 +620,7 @@ def run_xpress(
         "mip_unbounded": "unbounded",
     }
 
-    if io_api is not None and (io_api != "lp"):
+    if io_api is not None and io_api not in ["lp", "mps"]:
         logger.warning(
             f"IO setting '{io_api}' not available for xpress solver. "
             "Falling back to `lp`."